requirements.txt
flask
azure-storage-blob
azure-identity
